+++ CHANGES.md
@@ -9,6 +9,8 @@
 * Support `opam_version` to distro calls to build and install
   the latest version of OPAM2-dev.
 * Add Alpine 3.4 and deprecate Raspbian 7.
+* Remove Ubuntu 15.10 from default build list now that 16.10 LTS
+  is available.
 
 1.6.0:
 * Add a more modern Git in CentOS 6 to make it work with OPAM